+++ pyobjc-framework-Metal/Lib/Metal/__init__.py
@@ -27,5 +27,4 @@ sys.modules["Metal"] = mod = objc.ObjCLazyModule(
     (_Metal, AppKit),
 )
 
-
 del sys.modules["Metal._metadata"]